// only-doc-vibe-main/src/shared/api/cms/cms-schema.ts
export interface components {
  schemas: {
// ...
    LegalPage: {
      id: number;
      documentId: string;
      internal_title: string;
      slug: string;
      seo: components["schemas"]["SeoSeoComponent"];
      content: string;
      /** Format: date-time */
      createdAt: string;
      /** Format: date-time */
      updatedAt: string;
      /** Format: date-time */
      publishedAt: string;
      createdBy: {
        id: number;
        documentId: string;
      };
      updatedBy: {
        id: number;
        documentId: string;
      };
      locale: string;
      localizations: {
          id: number;
          documentId: string;
          internal_title: string;
          slug: string;
          seo: components["schemas"]["SeoSeoComponent"];
          content: string;
          /** Format: date-time */
          createdAt: string;
          /** Format: date-time */
          updatedAt: string;
          /** Format: date-time */
          publishedAt: string;
          createdBy: {
            id: number;
            documentId: string;
          };
          updatedBy: {
            id: number;
            documentId: string;
          };
          locale: string;
          localizations: {
              id: number;
              documentId: string;
            }[];
        }[];
    };
// ...
  };
// ...
}